Refactor API class into it's own model
Code meant for the end user and internal definitions became hard to
distinguish. Therefore Game has gotte its own file for the class in
which all API methods will be defined. The ui/ package will now only
define internal representations of the UI while the model/ package
contains contains the exposed API.
